+++ src/test/java/org/springframework/data/r2dbc/core/MySqlDatabaseClientIntegrationTests.java
@@ -36,7 +36,7 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.convert.ReadingConverter;
 import org.springframework.data.convert.WritingConverter;
 import org.springframework.data.r2dbc.dialect.MySqlDialect;
-import org.springframework.data.r2dbc.query.Criteria;
+import org.springframework.data.relational.core.query.Criteria;
 import org.springframework.data.r2dbc.testing.ExternalDatabase;
 import org.springframework.data.r2dbc.testing.MySqlTestSupport;
 import org.springframework.data.relational.core.mapping.Table;
@@ -46,6 +46,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
  * Integration tests for {@link DatabaseClient} against MySQL.
  *
  * @author Mark Paluch
+ * @author Mingyuan Wu
  */
 public class MySqlDatabaseClientIntegrationTests extends AbstractDatabaseClientIntegrationTests {
